fixed dependency in expression, sorting, grouping (#653) added regression to test 119
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@2578 406a0c4d-033a-0410-8de8-e80135713968

+++ src/yysphinxexpr.h
@@ -80,7 +80,7 @@
 typedef union YYSTYPE {
 	int64_t			iConst;			// constant value
 	float			fConst;			// constant value
-	int				iAttrLocator;	// attribute locator (rowitem for int/float; offset+size for bits)
+	uint64_t		iAttrLocator;	// attribute locator (rowitem for int/float; offset+size for bits)
 	int				iFunc;			// function id
 	int				iNode;			// node index
 } YYSTYPE;
